building up initialisation kernel

+++ apps/test_kernel.h
@@ -5,8 +5,8 @@ void test_kernel(double **value) {
         **value = 1.890567;
 }
 
-void test_kernel2(int **value) {
-        printf("%d ", (int)**value);
+void test_kernel2(double **value) {
+        printf("%lf ", (double)**value);
 }
 
 void test_kernel3(double **value, double **value_d) {
